CCDXParticleSystemQuad - memory corruption and potential leak fixes
CCDXParticleSystemQuad is used as a static, so it is initialized just
once for multiple CCParticleSystemQuads. As m_uMaxTotalParticles was
almost useless - some internal buffers have been initialized just
once (by CCDXParticleSystemQuad::initVertexAndIndexBuffer), so there
was a memory corruption in case if a small buffer (e.g. for 20 particles)
has been initialized originally, and later another CCParticleQuad with
more particles (but less than 100, for example 50) tried to be handled -
CCDXParticleSystemQuad::RenderVertexBuffer corruptes memory by rewriting
the originally allocated buffer with more data.
Consequences - random crashes, display flickering and so on. E.g. it was
possible that some CCParticleSystemQuads were not shown at all, some
are shown particularly (as less indices were provided than needed).

Such behaviour is fixed by this commit. m_uMaxTotalParticles is used now
as the currently allocated buffer size, and buffers are dynamically
reallocated when it is needed. No indexes are lost, no memory corruption.

Also a potential memory leak is fixed - if there is some problem
with mapping the buffer at CCDXParticleSystemQuad::RenderVertexBuffer