Text drawing - a partial fix with active ResolutionScale
In case if ResolutionScale is not equal 100 - the text metrics
are calculated incorrectly, also too big font is tried to be used.
Fix is partial - as seems there is a bug on DirectWrite side (?).
It does not handle alignment properly, so the bitmap with text
is clipping it really. So as initial version, to see something -
alignment is not supported for such displays (with ResolutionScale).
Full investigation should be done later.